Refactor for ubuntu and apt

+++ installers/nerdfont.sh
@@ -1,6 +1,5 @@
 #/bin/bash -ex
 
-
 # https://www.nerdfonts.com
 
 
@@ -12,10 +11,15 @@ case `uname` in
     ;;
     Linux)
         # commands for Linux go here
-        curl --location --output ~/downloads/FiraCode.zip https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/FiraCode.zip
-        unzip -o ~/downloads/FiraCode.zip -d ~/.fonts
-        rm ~/downloads/FiraCode.zip
-        fc-cache -fv
+        if ! compgen -G "${HOME}/.fonts/Fira*" > /dev/null; then
+            if command -v fc-cache &> /dev/null; then
+                # Ubuntu
+                curl --location --output ~/downloads/FiraCode.zip https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/FiraCode.zip
+                unzip -o ~/downloads/FiraCode.zip -d ~/.fonts
+                rm ~/downloads/FiraCode.zip
+                fc-cache -fv
+            fi
+        fi
     ;;
     FreeBSD)
         # commands for FreeBSD go here
